Renaming contentful file
