+++ pages/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Head from 'next/head'
-import { fetchContentfulEntries } from '../utils/contentfulPosts'
+import { fetchContentfulEntries } from '../utils/contentful'
 import Header from '../components/Header/Header'
 import Footer from '../components/Footer/Footer'
 import Shader from '../components/Shader/Shader'
